back/requirements.txt
aiohttp==3.6.2
amqp==2.5.2
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
billiard==3.6.1.0
boto3==1.12.30
botocore==1.15.30
cachetools==4.0.0
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
Django==3.0.7
django-anymail==7.1.0
django-axes==5.4.1
django-cors-headers==3.2.0
django-environ==0.4.5
django-fernet-fields==0.6
django-redis==4.11.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
docutils==0.15.2
fuzzywuzzy==0.18.0
gunicorn==20.0.4
google-auth==1.11.2
google-auth-oauthlib==0.4.1
idna==2.9
jmespath==0.9.5
kombu==4.6.7
more-itertools==8.0.2
multidict==4.7.5
nltk==3.4.5
oauthlib==3.1.0
packaging==19.2
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.3.2
pytest-django==3.7.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
sentry-sdk==0.14.4
six==1.13.0
slackclient==2.5.0
sparkpost==1.3.6
sqlparse==0.3.0
twilio==6.42.0
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.7
whitenoise==5.1.0
yarl==1.4.2